requirements.txt
pytorch
numpy
underthesea


